+++ rotkehlchen/chain/ethereum/modules/uniswap/utils.py
@@ -7,13 +7,19 @@ from typing import TYPE_CHECKING, Any, Callable, Optional
 import requests
 from web3.types import BlockIdentifier
 
+from rotkehlchen.accounting.structures.base import HistoryBaseEntry
+from rotkehlchen.accounting.structures.types import HistoryEventSubType, HistoryEventType
 from rotkehlchen.assets.asset import EvmToken
 from rotkehlchen.chain.ethereum.defi.zerionsdk import ZERION_ADAPTER_ADDRESS
 from rotkehlchen.chain.ethereum.interfaces.ammswap.types import LiquidityPool
 from rotkehlchen.chain.ethereum.interfaces.ammswap.utils import _decode_result
+from rotkehlchen.chain.ethereum.modules.uniswap.constants import CPT_UNISWAP_V2, CPT_UNISWAP_V3
+from rotkehlchen.chain.ethereum.utils import asset_normalized_value
 from rotkehlchen.chain.evm.contracts import EvmContract
+from rotkehlchen.chain.evm.decoding.utils import maybe_reshuffle_events
 from rotkehlchen.chain.evm.types import WeightedNode
 from rotkehlchen.constants import ZERO
+from rotkehlchen.constants.assets import A_ETH
 from rotkehlchen.constants.misc import ONE
 from rotkehlchen.constants.resolver import ethaddress_to_identifier
 from rotkehlchen.constants.timing import DEFAULT_TIMEOUT_TUPLE
@@ -235,3 +241,85 @@ def find_uniswap_v2_lp_price(
     numerator = (token0_supply * token0_price + token1_supply * token1_price)
     share_value = numerator / total_supply
     return Price(share_value)
+
+
+def decode_basic_uniswap_info(
+        amount_sent: int,
+        amount_received: int,
+        decoded_events: list[HistoryBaseEntry],
+        counterparty: str,
+        notify_user: Callable[[HistoryBaseEntry, str], None],
+) -> None:
+    """
+    Check last three events and if they are related to the swap, label them as such.
+    We check three events because potential events are: spend, (optionally) approval, receive.
+    Earlier events are not related to the current swap.
+    """
+    spend_event, approval_event, receive_event = None, None, None
+    for event in reversed(decoded_events):
+        try:
+            crypto_asset = event.asset.resolve_to_crypto_asset()
+        except (UnknownAsset, WrongAssetType):
+            notify_user(event, counterparty)
+            return
+
+        if (
+            event.event_type == HistoryEventType.INFORMATIONAL and
+            event.event_subtype == HistoryEventSubType.APPROVE and
+            approval_event is None
+        ):
+            approval_event = event
+        elif (
+            event.balance.amount == asset_normalized_value(amount=amount_sent, asset=crypto_asset) and  # noqa: E501
+            event.event_type == HistoryEventType.SPEND and
+            # don't touch ETH since there may be multiple ETH transfers
+            # and they are better handled by the aggregator decoder.
+            event.asset != A_ETH and
+            spend_event is None
+        ):
+            event.event_type = HistoryEventType.TRADE
+            event.event_subtype = HistoryEventSubType.SPEND
+            event.counterparty = counterparty
+            event.notes = f'Swap {event.balance.amount} {crypto_asset.symbol} in {counterparty}'  # noqa: E501
+            spend_event = event
+        elif (
+            event.balance.amount == asset_normalized_value(amount=amount_received, asset=crypto_asset) and  # noqa: E501
+            event.event_type == HistoryEventType.RECEIVE and
+            event.asset != A_ETH and
+            receive_event is None
+        ):
+            event.event_type = HistoryEventType.TRADE
+            event.event_subtype = HistoryEventSubType.RECEIVE
+            event.counterparty = counterparty
+            event.notes = f'Receive {event.balance.amount} {crypto_asset.symbol} as a result of a {counterparty} swap'  # noqa: E501
+            receive_event = event
+        elif (
+            event.counterparty in {CPT_UNISWAP_V2, CPT_UNISWAP_V3} and
+            event.event_type == HistoryEventType.TRADE
+        ):
+            # The structure of swaps is the following:
+            # 1.1 Optional approval event
+            # 1.2 Optional spend event
+            # 1.3 Optional receive event
+            # 1.4 SWAP_SIGNATURE event
+            # 2.1 Optional approval event
+            # 2.2 Optional spend event
+            # 2.3 Optional receive event
+            # 2.4 SWAP_SIGNATURE event
+            # etc.
+            # So if we are at SWAP_SIGNATURE № 2 then all events that are before SWAP_SIGNATURE № 1
+            # should have already been decoded, have counterparty set and have type Trade.
+            break
+        else:
+            # If what described in the comment above is not met then it is an error.
+            log.error(
+                f'Found unexpected event {event.serialize()} during decoding a uniswap swap '
+                f'in transaction {event.event_identifier.hex()}',
+            )
+            break
+
+    # Make sure that the approval event is NOT between the spend and receive events.
+    if spend_event is not None and approval_event is not None:
+        maybe_reshuffle_events(out_event=approval_event, in_event=spend_event)
+    if spend_event is not None and receive_event is not None:
+        maybe_reshuffle_events(out_event=spend_event, in_event=receive_event)